+++ app/candidate/profile/components/event-box-comps/event-box.tsx
