+++ app/candidate/profile/components/event-box-comps/event-card.tsx
