+++ app/candidate/profile/components/personal-info-comps/status-bar.tsx
